backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0
seaborn==0.12.2

# API和验证
pydantic>=2.0.3

# 其他工具
requests==2.31.0
joblib==1.3.1
beautifulsoup4>=4.12.0
python-whois>=0.8.0

# 日志
loguru==0.7.0

# 配置管理
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0

# 系统监控
psutil>=5.9.0